refactor: separate App component to use custom hook for WebSocket connection management
